iot_proj_final_fin/requirements.txt
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.1.0
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
protobuf==4.21.12
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
